Got the test to pass. My test that I created was extremely robust and helpful which helped me spot two bugs.
1. The date time was not be carried over to new records
2. it was filling in np.nan objects into the database which is less than ideal.

I fixed both of those things and all the tests passed all-right. I tested the front-end manually and it seemed to work well so far. I have yet to try the 'big file,' on it.

I also need to accept the other type of file (what they currently use to store data.), but that is a job for tomorrow.